Rename: Option enums 폴더로 이동
