+++ src/main/java/com/knu/KnowcKKnowcK/config/ChatGptConfig.java
@@ -2,10 +2,8 @@ package com.knu.KnowcKKnowcK.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
 import org.springframework.web.reactive.function.client.WebClient;
 
 @Configuration
